+++ container_service_extension/service.py
@@ -24,6 +24,8 @@ import container_service_extension.compute_policy_manager \
 from container_service_extension.config_validator import get_validated_config
 from container_service_extension.configure_cse import check_cse_installation
 from container_service_extension.consumer import MessageConsumer
+from container_service_extension.def_modules.models import DefEntityType
+from container_service_extension.def_modules.models import DefInterface
 import container_service_extension.def_modules.schema_svc as def_schema_svc
 import container_service_extension.def_modules.utils as def_utils
 from container_service_extension.def_modules.utils import raise_error_if_def_not_supported # noqa: E501
@@ -93,8 +95,8 @@ class Service(object, metaclass=Singleton):
         self.threads = []
         self.pks_cache = None
         self._state = ServerState.STOPPED
-        self._nativeInterface: def_utils.DefInterface = None
-        self._nativeEntityType: def_utils.DefEntityType = None
+        self._nativeInterface: DefInterface = None
+        self._nativeEntityType: DefEntityType = None
 
     def get_service_config(self):
         return self.config
